backend/requirements.txt
annotated-types
anyio
Brotli
certifi
cffi
charset-normalizer
click
colorama
distlib
dnspython
fastapi>=0.100.0
filelock
h11
idna
mutagen
pip
platformdirs
pycparser
pycryptodomex
pydantic>=2.0.0
pydantic_core
pymongo>=4.5.0
PyNaCl
pypng
python-dotenv
python-multipart
qrcode
requests
sniffio
starlette
typing_extensions
typing-inspection
urllib3
uvicorn>=0.23.0
virtualenv
websockets
yt-dlp
